added a generic component for crud operations (CrudTable)
- added CrudTable as a way to encapsulate all the code that handles the interaction of the index page with the store and passes it to the table.
- removed all code but the layout in index.ts for both users and testItems.
