+++ src/PackingListApp/ClientApp/src/pages/User/body.tsx
@@ -2,9 +2,13 @@
 
 // Modal for creation of new user
 export const NewUserView = (props: any) => {
-    return <div></div>
+    return (
+        <div>
+
+        </div>
+    )
 }
 
-export const UserFormBody = (props : any) => {
+export const UserFormBody = (props: any) => {
     return <div></div>
 }
